Created storage module for storage related stuff
